Give jenkins port directly
If we just give the jenkins URL, we are redirected with a 307, and
authentication headers are stripped. This is deliberate, as if an
attacker managed to redirect us to a different site we'd be giving
authentication data to them.

A more complete fix would have to check whether the redirect is trusted
and then re-try manually.